Move tests into cdm-misc. Make jenkins cdm-test tests all pass.
