+++ cdm-test/src/test/java/ucar/nc2/TestSequence.java
@@ -50,40 +50,13 @@ import java.util.Arrays;
 import java.util.List;
 
 /**
- * Test Sequences constructed when reading NLDN and BUFR datasets.
+ * Test Sequences constructed when reading BUFR datasets.
  *
  * @author caron
  * @since Nov 10, 2009
  */
 public class TestSequence {
   private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-
-  @Test
-  @Category(NeedsCdmUnitTest.class)
-  public void testRead() throws IOException {
-    try (NetcdfFile ncfile = NetcdfFiles.open(TestDir.cdmUnitTestDir + "ft/point/200929100.ingest")) {
-      Sequence record = (Sequence) ncfile.findVariable("record");
-
-      List<String> expectedMemberNames = Arrays.asList("tsec", "nsec", "lat", "lon", "sgnl", "mult", "fill",
-          "majorAxis", "eccent", "ellipseAngle", "chisqr");
-      Assert.assertEquals(Sets.newHashSet(expectedMemberNames), Sets.newHashSet(record.getVariableNames()));
-
-      try (StructureDataIterator iter = record.getStructureIterator()) {
-        int recordCount = 0;
-        while (iter.hasNext()) {
-          StructureData data = iter.next();
-
-          // Assert that a single value from the first record equals an expected value.
-          // Kinda lazy, but checking all values would be impractical.
-          if (recordCount++ == 0) {
-            Assert.assertEquals(-700, data.getScalarShort("sgnl"));
-          }
-        }
-
-        Assert.assertEquals(1165, recordCount);
-      }
-    }
-  }
 
   @Test
   @Category(NeedsCdmUnitTest.class)
