+++ cdm-test/src/test/java/ucar/nc2/ft/point/TestCfDocDsgExamples.java
@@ -11,6 +11,7 @@ import ucar.nc2.constants.FeatureType;
 import ucar.nc2.dataset.NetcdfDatasets;
 import ucar.nc2.write.NetcdfCopier;
 import ucar.nc2.write.NetcdfFormatWriter;
+import ucar.unidata.util.test.CheckPointFeatureDataset;
 import ucar.unidata.util.test.TestDir;
 import java.io.File;
 import java.io.FilenameFilter;
@@ -124,7 +125,8 @@ public class TestCfDocDsgExamples {
 
   @Test
   public void checkPointDataset() throws IOException {
-    Assert.assertEquals("npoints", countExpected, TestPointDatasets.checkPointFeatureDataset(location, ftype, show));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, ftype, show);
+    Assert.assertEquals("npoints", countExpected, checker.check());
   }
 
   /////////////////////////////////////////////////////////////////////////////////////////////////////////////
