+++ cdm-test/src/test/java/ucar/nc2/ft/point/TestMiscPointFeature.java
@@ -57,6 +57,7 @@ import ucar.nc2.time.CalendarDateRange;
 import ucar.unidata.geoloc.LatLonPoint;
 import ucar.unidata.geoloc.LatLonRect;
 import ucar.unidata.geoloc.Station;
+import ucar.unidata.util.test.CheckPointFeatureDataset;
 import ucar.unidata.util.test.category.NeedsCdmUnitTest;
 import ucar.unidata.util.test.TestDir;
 
@@ -67,59 +68,61 @@ public class TestMiscPointFeature {
   @Test
   public void testProblemProfile() throws IOException {
     String location = TestDir.cdmLocalFromTestDataDir + "point/profileMultidimZJoin.ncml";
-    Assert.assertEquals("npoints", 50, TestPointDatasets.checkPointFeatureDataset(location, FeatureType.PROFILE, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.PROFILE, true);
+    Assert.assertEquals("npoints", 50, checker.check());
   }
 
   @Test
   @Category(NeedsCdmUnitTest.class)
   public void testProblemTraj() throws IOException {
     String location = TestDir.cdmUnitTestDir + "ft/trajectory/cosmic/wetPrf_C005.2007.294.16.22.G17_0001.0002_nc";
-    Assert.assertEquals("npoints", 383,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.TRAJECTORY, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.TRAJECTORY, true);
+    Assert.assertEquals("npoints", 383, checker.check());
   }
 
   @Test
   public void testProblemStation() throws IOException {
     String location = TestDir.cdmLocalFromTestDataDir + "cfDocDsgExamples/H.2.4.1.ncml";
-    Assert.assertEquals("npoints", 100,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.STATION, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.STATION, true);
+    Assert.assertEquals("npoints", 100, checker.check());
   }
 
   @Test
   @Ignore("Dont support multiple lat/lon coordinates for now")
   public void testProblemStationWithPreciseCoords() throws IOException {
     String location = TestDir.cdmLocalFromTestDataDir + "cfDocDsgExamples/H.2.3.2.ncml";
-    Assert.assertEquals("npoints", 100,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.STATION, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.STATION, true);
+    Assert.assertEquals("npoints", 100, checker.check());
   }
 
   @Test
   @Category(NeedsCdmUnitTest.class)
   public void testGempakStationProfile() throws IOException {
     String location = TestDir.cdmUnitTestDir + "ft/sounding/gempak/19580807_upa.ncml";
-    Assert.assertEquals("npoints", 8769,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.STATION_PROFILE, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.STATION_PROFILE, true);
+    Assert.assertEquals("npoints", 8769, checker.check());
   }
 
   @Test
   public void testProblemStationProfile() throws IOException {
     String location = TestDir.cdmLocalFromTestDataDir + "point/stationProfileSingle.ncml";
-    Assert.assertEquals("npoints", 9,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.STATION_PROFILE, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.STATION_PROFILE, true);
+    Assert.assertEquals("npoints", 9, checker.check());
   }
 
   @Test
   public void testProblemSection() throws IOException {
     String location = TestDir.cdmLocalFromTestDataDir + "cfDocDsgExamples/H.6.3.1.ncml";
-    Assert.assertEquals("npoints", 145,
-        TestPointDatasets.checkPointFeatureDataset(location, FeatureType.TRAJECTORY_PROFILE, true));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.TRAJECTORY_PROFILE, true);
+    Assert.assertEquals("npoints", 145, checker.check());
   }
 
   @Test
   @Category(NeedsCdmUnitTest.class)
   public void testProblem3() throws IOException {
     String location = TestDir.cdmUnitTestDir + "ft/stationProfile/PROFILER_RASS_01hr_20091027_1500.nc";
-    assert 198 == TestPointDatasets.checkPointFeatureDataset(location, FeatureType.STATION_PROFILE, true);
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, FeatureType.STATION_PROFILE, true);
+    Assert.assertEquals("npoints", 198, checker.check());
   }
 
   @Test
