+++ cdm-test/src/test/java/ucar/nc2/ft/point/TestPreCFpointDatasets.java
@@ -12,6 +12,7 @@ import org.junit.runners.Parameterized;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ucar.nc2.constants.FeatureType;
+import ucar.unidata.util.test.CheckPointFeatureDataset;
 import ucar.unidata.util.test.TestDir;
 import java.io.File;
 import java.io.FileFilter;
@@ -52,7 +53,8 @@ public class TestPreCFpointDatasets {
 
   @Test
   public void checkPointDataset() throws IOException {
-    Assert.assertTrue("npoints", 0 < TestPointDatasets.checkPointFeatureDataset(location, ftype, show));
+    CheckPointFeatureDataset checker = new CheckPointFeatureDataset(location, ftype, show);
+    Assert.assertTrue("npoints", 0 < checker.check());
   }
 
 }
