+++ cdm-test/src/test/java/ucar/nc2/ncml/TestNcmlWriteAndCompareShared.java
@@ -2,7 +2,6 @@ package ucar.nc2.ncml;
 
 import org.apache.commons.io.filefilter.NotFileFilter;
 import org.apache.commons.io.filefilter.SuffixFileFilter;
-import org.jdom2.Element;
 import org.junit.*;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.TemporaryFolder;
@@ -10,20 +9,14 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ucar.nc2.NetcdfFile;
 import ucar.nc2.dataset.DatasetUrl;
-import ucar.nc2.dataset.NetcdfDataset;
-import ucar.nc2.dataset.NetcdfDatasets;
-import ucar.nc2.util.CompareNetcdf2;
-import ucar.nc2.write.NcmlWriter;
+import ucar.unidata.util.test.CompareNcml;
 import ucar.unidata.util.test.TestDir;
 import ucar.unidata.util.test.category.NeedsCdmUnitTest;
-import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
 import java.util.ArrayList;
-import java.util.Formatter;
 import java.util.List;
 
 /**
@@ -67,7 +60,6 @@ public class TestNcmlWriteAndCompareShared {
 
     result.add(new Object[] {datadir + "formats/gini/n0r_20041013_1852-compress", false}); //
     result.add(new Object[] {datadir + "formats/gini/ntp_20041206_2154", true}); //
-    result.add(new Object[] {datadir + "formats/dmsp/F14200307192230.n.OIS", false}); //
 
     result.add(new Object[] {datadir + "formats/nexrad/level2/6500KHGX20000610_000110", false});
     result.add(new Object[] {datadir + "formats/nexrad/level2/Level2_KYUX_20060527_2335.ar2v", true});
@@ -97,174 +89,18 @@ public class TestNcmlWriteAndCompareShared {
   }
 
   /////////////////////////////////////////////////////////////
-  boolean showFiles = true;
-  boolean compareData = false;
+  private final boolean showFiles = true;
+  private final boolean compareData;
+  private final DatasetUrl durl;
 
   public TestNcmlWriteAndCompareShared(String location, boolean compareData) throws IOException {
     this.durl = DatasetUrl.findDatasetUrl(location);
     this.compareData = compareData;
   }
 
-  // String location;
-  DatasetUrl durl;
-
-  int fail = 0;
-  int success = 0;
-
   @Test
   public void compareNcML() throws IOException {
-    compareNcML(true, true, true);
-    compareNcML(true, false, false);
-    compareNcML(false, true, false);
-    compareNcML(false, false, true);
-    compareNcML(false, false, false);
-  }
-
-  private void compareNcML(boolean useRecords, boolean explicit, boolean openDataset) throws IOException {
-    if (compareData)
-      useRecords = false;
-
-    if (showFiles) {
-      System.out.println("-----------");
-      System.out.println("  input filename= " + durl.trueurl);
-    }
-
-    NetcdfFile org;
-    Object iospMessage = useRecords ? NetcdfFile.IOSP_MESSAGE_ADD_RECORD_STRUCTURE : null;
-    if (openDataset)
-      org = NetcdfDatasets.openDataset(durl, null, -1, null, iospMessage);
-    else
-      org = NetcdfDatasets.acquireFile(null, null, durl, -1, null, iospMessage);
-
-    // create a file and write it out
-    int pos = durl.trueurl.lastIndexOf("/");
-    String ncmlOut = tempFolder.newFile().getAbsolutePath();
-    if (showFiles)
-      System.out.println(" output filename= " + ncmlOut);
-
-    try {
-      NcmlWriter ncmlWriter = new NcmlWriter();
-      Element netcdfElement;
-
-      if (explicit) {
-        netcdfElement = ncmlWriter.makeExplicitNetcdfElement(org, null);
-      } else {
-        netcdfElement = ncmlWriter.makeNetcdfElement(org, null);
-      }
-
-      ncmlWriter.writeToFile(netcdfElement, new File(ncmlOut));
-    } catch (IOException ioe) {
-      // ioe.printStackTrace();
-      assert false : ioe.getMessage();
-    }
-
-    // read it back in
-    NetcdfFile copy;
-    DatasetUrl durlcopy = DatasetUrl.findDatasetUrl(ncmlOut);
-    if (openDataset)
-      copy = NetcdfDatasets.openDataset(durlcopy, null, -1, null, iospMessage);
-    else
-      copy = NetcdfDatasets.acquireFile(null, null, durlcopy, -1, null, iospMessage);
-
-    try {
-      Formatter f = new Formatter();
-      CompareNetcdf2 mind = new CompareNetcdf2(f, false, false, compareData);
-      boolean ok = mind.compare(org, copy, new CompareNetcdf2.Netcdf4ObjectFilter());
-      if (!ok) {
-        fail++;
-        System.out.printf("--Compare %s, useRecords=%s explicit=%s openDataset=%s compareData=%s %n", durl.trueurl,
-            useRecords, explicit, openDataset, compareData);
-        System.out.printf("  %s%n", f);
-      } else {
-        System.out.printf("--Compare %s is OK (useRecords=%s explicit=%s openDataset=%s compareData=%s)%n",
-            durl.trueurl, useRecords, explicit, openDataset, compareData);
-        success++;
-      }
-      Assert.assertTrue(durl.trueurl, ok);
-    } finally {
-      org.close();
-      copy.close();
-    }
-  }
-
-  // @Test
-  public void compareNcMLold() throws IOException {
-    compareNcMLold(true, true, true);
-    // compareNcMLold(true, false, false);
-    // compareNcMLold(false, true, false);
-    // compareNcMLold(false, false, true);
-    // compareNcMLold(false, false, false);
-  }
-
-  public void compareNcMLold(boolean useRecords, boolean explicit, boolean openDataset) throws IOException {
-    if (compareData)
-      useRecords = false;
-
-    if (showFiles) {
-      System.out.println("-----------");
-      System.out.println("  input filename= " + durl.trueurl);
-    }
-
-    NetcdfFile org;
-    if (openDataset)
-      org = NetcdfDataset.openDataset(durl.trueurl, false, null);
-    else
-      org = NetcdfDataset.acquireFile(durl, null);
-
-    if (useRecords)
-      org.sendIospMessage(NetcdfFile.IOSP_MESSAGE_ADD_RECORD_STRUCTURE);
-
-    // create a file and write it out
-    int pos = durl.trueurl.lastIndexOf("/");
-    String ncmlOut = tempFolder.newFile().getAbsolutePath();
-    if (showFiles)
-      System.out.println(" output filename= " + ncmlOut);
-
-    try {
-      NcmlWriter ncmlWriter = new NcmlWriter();
-      Element netcdfElement;
-
-      if (explicit) {
-        netcdfElement = ncmlWriter.makeExplicitNetcdfElement(org, null);
-      } else {
-        netcdfElement = ncmlWriter.makeNetcdfElement(org, null);
-      }
-
-      ncmlWriter.writeToFile(netcdfElement, new File(ncmlOut));
-    } catch (IOException ioe) {
-      // ioe.printStackTrace();
-      assert false : ioe.getMessage();
-    }
-
-    // read it back in
-    NetcdfFile copy;
-    if (openDataset)
-      copy = NetcdfDataset.openDataset(ncmlOut, false, null);
-    else
-      copy = NetcdfDataset.acquireFile(DatasetUrl.findDatasetUrl(ncmlOut), null);
-
-    if (useRecords)
-      copy.sendIospMessage(NetcdfFile.IOSP_MESSAGE_ADD_RECORD_STRUCTURE);
-
-    try {
-      Formatter f = new Formatter();
-      CompareNetcdf2 mind = new CompareNetcdf2(f, false, false, compareData);
-      boolean ok = mind.compare(org, copy, new CompareNetcdf2.Netcdf4ObjectFilter());
-      if (!ok) {
-        fail++;
-        System.out.printf("--Compare %s, useRecords=%s explicit=%s openDataset=%s compareData=%s %n", durl.trueurl,
-            useRecords, explicit, openDataset, compareData);
-        System.out.printf("  %s%n", f);
-      } else {
-        System.out.printf("--Compare %s is OK (useRecords=%s explicit=%s openDataset=%s compareData=%s)%n",
-            durl.trueurl, useRecords, explicit, openDataset, compareData);
-        success++;
-      }
-      Assert.assertTrue(durl.trueurl, ok);
-    } finally {
-      org.close();
-      copy.close();
-    }
+    new CompareNcml(tempFolder, durl, compareData, showFiles);
   }
 
 }
